Module depy::parsing::parse_json

Json reading and transforming it to Manifest

Functions

expand_vars

Expands variables that are needed in env_set and env_add_path

get_env_paths

Returns all env_add_path attributes from a manifest in the form of a Vec<EnvVar>

get_env_variables

Returns all env_set attributes from a manifest in the form of a Vec<EnvVar>

get_version

Returns the version of a manifest

parse_arch

Gets the current system arch and returns it in a format understood by scoop

query_bin

Checks all bins in a manifest for a certain query

read_json_file

Parses a json file from a path and returns the contents as a serde_json::Value