Module depy::shell::bucket

Bucket related shell operations

Functions

add_bucket

Adds a bucket to the depy/scoop instalation

clean_buckets

Deletes all buckets installed in the depy/scoop instalation

list_buckets

Returns Vec<(name, url)>

parse_bucket

Returns the url of a bucket in a githubusercontent raw form

query_all_buckets

Finds a package matching the query in any buckets in the depy/scoop/buckets

remove_bucket

Remove a bucket to the depy/scoop instalation