Module depy::shell::dir

Directory reading/modifing related shell operations

Functions

clear_directory

Deletes everything inside a directory

get_depy_scoop_location

Returns the location of the depy installation directory

get_scoop_dir_location

Returns the location of the main scoop installation directory

get_version_location

Returns the location of a specific version of an app

init_depy_dir

Creates all directories necessary for depy to function under the %userprofile%/depy directory